rss package passes tests
