+++ services/rss/pyproject.toml
@@ -9,7 +9,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "rss"
 version = "1.7.0"
-description = "API for the rss website"
+description = "Package for creating RSS feeds for earthquakes in CAP format."
 readme = "README.md"
 authors = [{ name = "Daniel Ibarrola", email = "daniel.ibarrola.sanchez@gmail.com" }]
 classifiers = [
@@ -20,26 +20,7 @@ keywords = ["earthquake", "alerts", "cap", "rss"]
 requires-python = ">=3.11"
 
 [project.optional-dependencies]
-dev = ["bumpver", "pytest", "pytest-mock", "flask-cors"]
+dev = ["pytest"]
 
 [project.urls]
 Homepage = "https://github.com/Daniel-Ibarrola/RSS.git"
-
-[project.scripts]
-cap = "rss.__main__:main"
-
-[tool.bumpver]
-current_version = "1.7.0"
-version_pattern = "MAJOR.MINOR.PATCH"
-commit_message = "bump version {old_version} -> {new_version}"
-tag_message = "{new_version}"
-tag_scope = "default"
-commit = true
-tag = true
-push = true
-
-[tool.bumpver.file_patterns]
-"pyproject.toml" = [
-    'current_version = "{version}"',
-    'version = "{version}"'
-]
