+++ services/rss/requirements.txt
@@ -1,0 +1,1 @@
+Babel==2.13.1
