+++ services/rss/src/rss.egg-info/PKG-INFO
@@ -1,0 +1,13 @@
+Metadata-Version: 2.1
+Name: rss
+Version: 1.7.0
+Summary: Package for creating RSS feeds for earthquakes in CAP format.
+Author-email: Daniel Ibarrola <daniel.ibarrola.sanchez@gmail.com>
+Project-URL: Homepage, https://github.com/Daniel-Ibarrola/RSS.git
+Keywords: earthquake,alerts,cap,rss
+Classifier: Programming Language :: Python
+Classifier: Programming Language :: Python :: 3
+Requires-Python: >=3.11
+Description-Content-Type: text/markdown
+Provides-Extra: dev
+Requires-Dist: pytest; extra == "dev"
