+++ services/rss/src/rss.egg-info/requires.txt
@@ -1,0 +1,3 @@
+
+[dev]
+pytest
